Create filetypedetect augroup in filetype.vim
Like the plugin file, filetype.vim is automatically sourced by Neovim,
but it is done much sooner than plugin files (it is sourced immediately
after the user's init file). It turns out that this ordering matters for
autocommand definitions: autocommands cannot specify any kind of
priority, so they run in the order they are defined in. The builtin
filetype.vim creates the filetypedetect autocommands early on, which
means users (and plugins) can use the buffer's filetype in their own
BufRead autocommands.

However, when the filetypedetect autocommand is loaded in the _plugin_
file, any BufRead autocommands defined _before_ that file is sourced are
not able to read the buffer's filetype.

Move the filetypedetect autocommand into filetype.vim to ensure it's
created as soon as possible. This also is more consistent with the
builtin filetype.vim and ensures there are fewer "gotchas" when using
filetype.nvim.